Change swagger entry to reflect millicents change; move unit methods to millicents.go
